refactor: 利用時間帯、時分を別ファイルに抽出
