+++ 会議室予約/Domain/予約.cs
@@ -19,6 +19,11 @@ namespace 会議室予約.Domain
             this.かいぎしつ = かいぎしつ;
             this.かいぎさんかよていしゃ = かいぎさんかよていしゃ;
         }
+
+        public 利用時間帯 りようじかんたい()
+        {
+            return りようきかん.りようじかんたい();
+        }
         
         // 操作・メソッド考える？
         // public 予約()
