+++ 会議室予約/Domain/利用可能時間帯.cs
