+++ 会議室予約/会議室予約.csproj
@@ -4,4 +4,8 @@
         <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <Folder Include="Domain\type" />
+    </ItemGroup>
+
 </Project>
